Stabilize API for upcoming release
This adopts a (fairly standard) approach for API/ABI stability and
versioning.

The reason for stabilizing the API is more due to the fact that
development has slowed considerably, and stabilizing it will facilitate
wider and more standardized distribution of the library.  Indeed, there
are probably a number of areas where the API could still be
significantly improved, but no such work is currently under way.  If and
when it does, we'll target it to a Version 2.0.

Signed-off-by: Bryce Harrington <bryce@bryceharrington.org>